% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.med.R
\name{draw.med}
\alias{draw.med}
\title{Simple Mediation Diagrams}
\usage{
draw.med(model1, model2, model3, y, x, m)
}
\arguments{
\item{model1}{The saved regression model of \code{y~x} model where X predicts Y.}

\item{model2}{The saved regression model of \code{m~x} model where X predicts M.}

\item{model3}{The saved regression model of \code{y~x+m} model where X and M predict Y.}
}
\description{
This function returns the code to create a simple mediation triangle,
as well as the basic diagram using the \code{diagram} library.
}
\examples{
draw.med(saved$model1, saved$model2, saved$model3)
Note in this example saved is the name of the model saved from mediation1.
You can include any form of model names that you have saved from \\code{lm()}.
You can also type \\code{draw.med} in your console to get this code and edit
the diagram parameters to your liking.
}
\keyword{bootstrapping,}
\keyword{data}
\keyword{diagram}
\keyword{mediation,}
\keyword{regression,}
\keyword{screening,}
