% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirectm1.R
\name{indirectm1}
\alias{indirectm1}
\title{indirectm1}
\usage{
indirectm1(formula2, formula3, df, random)
}
\arguments{
\item{formula2}{The formula for mediation for the a path, usually `m ~ x`. Can also include covariates and will be `eq2` if the `createformula()` function is used.}

\item{formula3}{The formula for mediation for the b path, usually `y ~ x + m`. Can also include covariates and will be `eq3` if the `createformula()` function is used.}

\item{df}{The dataframe where the columns from the formula can be found.}

\item{random}{This variable is used to denote the data frame will be randomize by row, as part of the `boot` library.}
}
\description{
This function runs a simple mediation model to calculate the indirect effect, which will be used for bootstrapping the confidence interval of the indirect effect.
}
\examples{
indirectm1("disp ~ mpg", "cyl ~ mpg + disp", mtcars)
}
\keyword{effect}
\keyword{indirect}
\keyword{mediation,}
\keyword{moderation,}
\keyword{regression,}
