% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirectmed.R
\name{indirectmed}
\alias{indirectmed}
\title{Indirect Effect for Mediation}
\usage{
indirectmed(formula2, formula3, x, m, df, random)
}
\arguments{
\item{formula2}{The formula for mediation for the a path, usually
\code{m ~ x}. Can also include covariates and will be \code{eq2}
if the \code{createformula()} function is used.}

\item{formula3}{The formula for mediation for the b path, usually
\code{y ~ x + m}. Can also include covariates and will be
\code{eq3} if the \code{createformula()} function is used.}

\item{x}{The column name for x in the data frame.}

\item{m}{The column name for m in the data frame.}

\item{df}{The dataframe where the columns from the formula can be found.}

\item{random}{This variable is used to denote the data frame will be
randomize by row, as part of the \code{boot} library.}
}
\description{
This function runs a simple mediation model to calculate
the indirect effect, which will be used for bootstrapping
the confidence interval of the indirect effect. This function
is used in conjunction with the \code{boot} function and is formatted to
work as a bootstrapped effect.
}
\examples{
indirectmed("disp ~ mpg", "cyl ~ mpg + disp", mtcars)
}
\keyword{effect}
\keyword{indirect}
\keyword{mediation,}
\keyword{moderation,}
\keyword{regression,}
