% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation2.R
\name{mediation2}
\alias{mediation2}
\title{Serial Mediation with Two Mediators}
\usage{
mediation2(y, x, m1, m2, cvs = NULL, df, with_out = T, nboot = 1000,
  conf_level = 0.95)
}
\arguments{
\item{y}{The dependent variable column name from your dataframe.}

\item{x}{The independent variable column name from your dataframe. This column
will be treated as X in mediation or moderation models, please see
diagrams online for examples.}

\item{m1}{The first mediator for your model.}

\item{m2}{The second mediator for your model.}

\item{cvs}{The covariates you would like to include in the model.
Use a \code{c()} concatenated vector to use multiple covariates.}

\item{df}{The dataframe where the columns from the formula can be found.
Note that only the columns used in the analysis will be data screened.}

\item{with_out}{A logical value where you want to keep the outliers in
model \code{TRUE} or exclude them from the model \code{FALSE}.}

\item{nboot}{A numeric value indicating the number of bootstraps you would like to complete.}

\item{conf_level}{A numeric value indicating the confidence interval width for the boostrapped confidence interval.}
}
\description{
This function runs a complete serial mediation analysis with two
mediator, similiar to model 6 in PROCESS by A. Hayes (2013).
As part of the output, you will find data screening,
all three models used in the traditional Baron and
Kenny (1986) steps, total/direct/indirect effects, the z-score and p-value
for the Aroian Sobel test, and the bootstrapped confidence interval
for the indirect effect.
}
\examples{
mediation2(y = "Q11", x = "Q151", m1 = "Q31", m2 = "Q41",
          cvs = c("Q121"), df = mediation2, nboot = 1000, with_out = T,
          conf_level = .95)
}
\keyword{bootstrapping}
\keyword{data}
\keyword{mediation,}
\keyword{regression,}
\keyword{screening,}
