% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation2_data.R
\docType{data}
\name{mediation2_data}
\alias{mediation2_data}
\title{Mediation 2 Data}
\format{A data frame of student evaluation ratings}
\usage{
data(mediation2_data)
}
\description{
Dataset for use in \code{\link{mediation2}} example in MeMoBootR.
These items were taken from student evaluations of teachers with
15 questions. ID represents a unique ID for each rating.
}
\keyword{datasets}
