% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation1.R
\name{moderation1}
\alias{moderation1}
\title{Two Way Moderation}
\usage{
moderation1(y, x, m, cvs = NULL, df, with_out = T)
}
\arguments{
\item{y}{The dependent variable column name from your dataframe.}

\item{x}{The independent variable column name from your dataframe. This column
will be treated as X in mediation or moderation models, please see
diagrams online for examples.}

\item{m}{The moderator for your model, as this model only includes one moderator.}

\item{cvs}{The covariates you would like to include in the model.
Use a \code{c()} concatenated vector to use multiple covariates.}

\item{df}{The dataframe where the columns from the formula can be found.
Note that only the columns used in the analysis will be data screened.}

\item{with_out}{A logical value where you want to keep the outliers in
model \code{TRUE} or exclude them from the model \code{FALSE}.}
}
\description{
This function runs a complete two way moderation analysis with one
moderator, similiar to model 1 in PROCESS by A. Hayes (2013).
As part of the output, you will find data screening,
the overall model, and the simple slopes for X at each level of the moderator.
X and M variables will be mean centered after data screening and
before analysis to control for multicollinearity unless they are categorical.
}
\examples{
states = as.data.frame(state.x77)
moderation1(y = "Income", x = "Illiteracy", m = "Murder",
          cvs = c("Population", "Area"), df = states)
}
\keyword{bootstrapping}
\keyword{data}
\keyword{mediation,}
\keyword{moderation,}
\keyword{regression,}
\keyword{screening,}
