% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation2.R
\name{moderation2}
\alias{moderation2}
\title{Doule Two-Way Moderation}
\usage{
moderation2(y, x, m1, m2, cvs = NULL, df, with_out = T)
}
\arguments{
\item{y}{The dependent variable column name from your dataframe.}

\item{x}{The independent variable column name from your dataframe. This column
will be treated as X in mediation or moderation models, please see
diagrams online for examples.}

\item{m1}{The first moderator for your model.}

\item{m2}{The second moderator for your model.}

\item{cvs}{The covariates you would like to include in the model.
Use a \code{c()} concatenated vector to use multiple covariates.}

\item{df}{The dataframe where the columns from the formula can be found.
Note that only the columns used in the analysis will be data screened.}

\item{with_out}{A logical value where you want to keep the outliers in
model \code{TRUE} or exclude them from the model \code{FALSE}.}
}
\description{
This function runs a complete double two-way moderation analysis with two
moderators, similiar to model 2 in PROCESS by A. Hayes (2013).
As part of the output, you will find data screening,
the overall model, and the simple slopes for X at each level of the moderator.
X and M variables will be mean centered after data screening and
before analysis to control for multicollinearity unless they are categorical.
}
\examples{
states = as.data.frame(state.x77)
moderation2(y = "Income", x = "Illiteracy", m1 = "Murder",
          m2 = "Population", df = states)
}
\keyword{bootstrapping}
\keyword{data}
\keyword{mediation,}
\keyword{moderation,}
\keyword{regression,}
\keyword{screening,}
