% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createformula.R
\name{createformula}
\alias{createformula}
\title{createformula}
\usage{
createformula(y, x, m = NULL, cvs = NULL)
}
\arguments{
\item{y}{The dependent variable column name from your dataframe.}

\item{x}{The independent variable column name from your dataframe. This column will be treated as X in mediation or moderation models, please see diagrams online for examples.}

\item{m}{The first mediator or moderator for your model. This function will incorporate multiple mediators/moderators later.}

\item{cvs}{The covariates you would like to include in the model. Use a `c()` concatenated vector to use multiple covariates.}
}
\description{
This function creates the formulas for `lm()` to run the regression models necessary for data screening and mediation or moderation models. 
The output includes a list of equations used for mediation, which will updated as moderation is added to the package.
}
\examples{
createformula(y = "cyl", x = "mpg", m = "disp", cvs = c("drat", "gear"))
}
\keyword{formulas}
\keyword{mediation,}
\keyword{moderation,}
\keyword{regression,}
