% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createformula.R
\name{createformula}
\alias{createformula}
\title{Create Regression Formulas}
\usage{
createformula(y, x, m, m2 = NULL, cvs = NULL, type)
}
\arguments{
\item{y}{The dependent variable column name from your dataframe.}

\item{x}{The independent variable column name from your dataframe. This column will be treated as X in mediation or moderation models, please see diagrams online for examples.}

\item{m}{The first mediator or moderator for your model.}

\item{m2}{The second mediator or moderator for your model.}

\item{cvs}{The covariates you would like to include in the model. Use a `c()` concatenated vector to use multiple covariates.}
}
\description{
This function creates the formulas for \code{lm()} to run the
regression models necessary for data screening and mediation
or moderation models.
}
\details{
The output currently includes a list of equations used for simple mediation,
which will updated as moderation is added to the package.
}
\examples{
createformula(y = "cyl", x = "mpg", m = "disp",
             cvs = c("drat", "gear"), type = "moderation1")
}
\keyword{formulas}
\keyword{mediation,}
\keyword{moderation,}
\keyword{regression,}
